+++ geoimagenet_ml/ml/impl.py
@@ -1092,10 +1092,10 @@ def create_batch_patches(annotations_meta,      # type: List[JSON]
 
     train_counter, test_counter = category_counter.split(train_test_ratio)
     train_test_splits = [("train", train_counter), ("test", test_counter)]
-    patches_crop = [(None, "raw")]
+    patches_crop = [(None, 1, "raw")]
     if isinstance(crop_fixed_size, (int, float)):
         update_func("fixed sized crops [{}] also selected for creation".format(crop_fixed_size), start_percent)
-        patches_crop.append((crop_fixed_size, "fixed"))
+        patches_crop.append((crop_fixed_size, None, "fixed"))
 
     last_progress_offset, progress_scale = 0, float(patch_percent - start_percent) / len(features)
     dataset_container.data["patches"] = list()
@@ -1142,8 +1142,8 @@ def create_batch_patches(annotations_meta,      # type: List[JSON]
                 DATASET_DATA_PATCH_FEATURE_KEY: feature.get("id"),
             })
 
-            for crop_size, crop_name in patches_crop:
-                crop, _, bbox = process_feature_crop(feature["geometry"], srs, raster_data, crop_size)
+            for crop_size, crop_mode, crop_name in patches_crop:
+                crop, inv_crop, bbox = process_feature_crop(feature["geometry"], srs, raster_data, crop_size, crop_mode)
                 if crop is not None:
                     if crop.ndim < 3 or crop.shape[2] != raster_data["band_count"]:
                         raise AssertionError("bad crop channel size")
@@ -1153,7 +1153,7 @@ def create_batch_patches(annotations_meta,      # type: List[JSON]
                     output_name = get_sane_name("{}_{}".format(feature["id"], crop_name), assert_invalid=False)
                     output_path = os.path.join(dataset_container.path, "{}.tif".format(output_name))
                     output_mask = os.path.join(dataset_container.path, "{}_mask.png".format(output_name))
-                    crop_image = Image.fromarray((255*crop.mask).astype(np.uint8))
+                    crop_image = Image.fromarray((255 * inv_crop.mask).astype(np.uint8)).convert("RGB")  # remove alpha
                     crop_image.save(output_mask)
                     if os.path.exists(output_path):
                         msg = "Output path [{}] already exists but is expected to not exist.".format(output_path)